\name{bdfunc.enrichment.human}
\alias{bdfunc.enrichment.human}
\docType{data}
\title{
BD-Func Enrichment File for Human Gene Symbols
}
\description{
Contains genes lists of paired up-regulated and down-regulated genes.  Gene Lists come from Gene Ontology (GO [1]) and MSigDB [2].

BD-Func [3] will compare the expression patterns for the up-regulated genes to the down-regulated genes.
}
\usage{data(bdfunc.enrichment.human)}

\source{
GO Gene Lists: http://www.geneontology.org/GO.downloads.annotations.shtml
MSigDB Gene Lists: http://www.broadinstitute.org/gsea/downloads.jsp
}
\references{
[1] Ashburner M, Ball CA, Blake JA, Botstein D, Butler H, Cherry JM,
 Davis AP, Dolinski K, Dwight SS, Eppig JT, Harris MA, Hill DP, Issel-Tarver L,
 Kasarskis A, Lewis S, Matese JC, Richardson JE, Ringwald M, Rubin GM,
 and Sherlock G.(2000).
Gene Ontology: tool for the unification of biology
\emph{Nat Genet},
25:25-29

[2] Liberzon A, Subramanian A, Pinchback R, Thorvaldsdottir H, Tamayo P, and Mesirov JP.(2011).
Molecular signatures database (MSigDB) 3.0.
\emph{Bioinformatics},
27:1739-1740.

[3] Warden CD, Kanaya N, Chen S, and Yuan Y-C. (2013).
BD-Func: A Streamlined Algorithm for Predicting Activation and Inhibition of Pathways.
\emph{peerJ},
1:e159
}
\keyword{datasets}
