\name{bdfunc.enrichment.mouse}
\alias{bdfunc.enrichment.mouse}
\docType{data}
\title{
BD-Func Enrichment File for Mouse Gene Symbols
}
\description{
Contains genes lists of paired up-regulated and down-regulated genes.  Gene Lists come from Gene Ontology (GO [1]).

BD-Func [2] will compare the expression patterns for the up-regulated genes to the down-regulated genes.
}
\usage{data(bdfunc.enrichment.mouse)}

\source{
GO Gene Lists: http://www.geneontology.org/GO.downloads.annotations.shtml
}
\references{
[1] Ashburner M, Ball CA, Blake JA, Botstein D, Butler H, Cherry JM,
 Davis AP, Dolinski K, Dwight SS, Eppig JT, Harris MA, Hill DP, Issel-Tarver L,
 Kasarskis A, Lewis S, Matese JC, Richardson JE, Ringwald M, Rubin GM,
 and Sherlock G.(2000).
Gene Ontology: tool for the unification of biology
\emph{Nat Genet},
25:25-29

[2] Warden CD, Kanaya N, Chen S, and Yuan Y-C. (2013).
BD-Func: A Streamlined Algorithm for Predicting Activation and Inhibition of Pathways.
\emph{peerJ},
1:e159
}
\keyword{datasets}
